Define media route protocols for media playback.
Added a new support library, android-support-media-protocol.jar,
which defines media route protocols that applications may use to
control media devices via media routes.

Adapted the Cast remote media player as the basis of a generic
media player protocol that may be supported by many different
types of devices.  The purpose here is similar in nature to
the Intent based protocols that were defined in the original
media router support library (CATEGORY_REMOTE_PLAYBACK) but
this new protocol is somewhat more comprehensive and will hopefully
be easier to use as well.

Change-Id: I69dfaca01fc710f3169bbf9aed798f607287b7ca