DO NOT MERGE Update RecyclerView scroll listener API
This change adds RecyclerView as a first parameter to the scroll related callbacks.

It also fxies a bug where scroll callback was being called w/ the intended scroll
amount instead of the real scroll amount.

I also changed it to be an abstract class instead of an interface to make future
changes easier.

Bug: 16054949
Change-Id: I99a1f91dcd442c1ac3ad1ed7dbf4598d1cee9c28

Conflicts:
	v17/leanback/src/android/support/v17/leanback/widget/DetailsOverviewRowPresenter.java
	v7/recyclerview/src/android/support/v7/widget/RecyclerView.java