Don't animate the ActionBarDrawerToggle drawer indicator for new apps
The new Toolbar-backed action bars provide enough space for navigation
that the drawer glyph has its own permanent button-like
position. Don't slide it back and forth like the Holo version if the
app is new and running on a new platform version. Also let the app
decide this manually.

Bug 15885014

Change-Id: I43b66b750d8db13ad4da646ee70fd39d8019c257