Update colorAccent values + fix precomputed drawables
The script I created uses ImageMagick to precompute
the resulting PNGs, but ImageMagick honors the alpha
value in the tint colors, which different to Android's
PorterDuff based Drawable tints.

I've also added a small optimization to the script
which reduces the amount of Drawables produced.
Previously if the dark and light variants used the
same tint color, we would still produce two PNGs.
We now only produce one.

Finally, I've added the script to the source folder
so that it can updated and used as necessary.

BUG: 16677464

Change-Id: I2370139129c1674c68c3b0e94c5f19f90a6d1472