Improve LinearLayoutManager's predictive layout handling
This CL updates LLM to select anchor child more carefully. Previously, it was
selecting a child whose position is withing State#itemCount bounds. Now it
chooses a visible child which is not removed. If cannot find such child,
prefers a child that is not removed but not within view  bounds. If it does
not exists either, falls back to a child whose position is within adapter
bounds.

Change-Id: I6aa759b17d80b5f08d4e874bda177501fa1acbe3