Stop settling sooner
Logic in ViewDragHelper.continueSettling() is wrong - when we've
reached our final resting place, we set a 'keepGoing' flag to be the
state of 'finished' on the scroller. This is the opposite of what it
should be and forces us to come back to this method one more time.
We should simply stop settling during that frame instead, since
we've just told the scroller to set the final values and to finish
the animation.

Issue #17108393 ViewDragHelper "keepGoing" condition assigned the wrong value after aborts

Change-Id: I9714bb1a0802bbadd6087ba2064fe23bd6cc34f0