Add a print finished callback to PrintHelper.
Applications that implement printing via the PrintHelper need to know
when printing is finished, i.e. the system got all the data for the
print job or the print job was cancelled, because they should not finish
the activity that initiated printing until the print interaction finishes.
Othwrwise, the connection between this activity and the print framework
will break and printing will fail.

bug:17367152

Change-Id: I9a25bc44e6dd0331fb146daf66ad81f39802ebdc