Skip removing the Fragment's view if it is being animated.
The animation is currently being canceled because the view is removed
from its parent immediately after starting the animation, so the
animation never plays.  Likewise, the state never advances because
the animation listener's onAnimationEnd() callback is not invoked.

This change defers removal of the mAnimatingAway view until the
animation finishes playing.

Cherry-pick of https://android-review.googlesource.com/#/c/107740/

Change-Id: I0db9cee5a13580a8a92b6869f174a1ae2ae22c18