DrawerLayout insets support for drawing under the status bar
When DrawerLayout is set to android:fitSystemWindows="true", it can
now intelligently apply insets to child views as well as draw
background protection for the status bar.

If a child view does not itself specify fitsSystemWindows=true, it
will have its margins altered to inset for the system window insets.
If the child does, it will have insets dispatched to it accordingly.

Bug 17547635

Change-Id: Ifa8e2a681ad4915421715870cb8be27df76e1132