DO NOT MERGE Set transitionName for rebind after transition
After transitionEnd,  people may rebind row to view, where
we may re-create the view (even though the view should be
in recycle pool unless app calls notifyChange() which invalidate
all the recycle pool).  So we should always setTransitionName()
on the view for a back transition.  Meanwhile we should clear
transitionName on existing view when rebind to a new view.

Added logs.

b/17918639

Change-Id: I204cf847114e5d379e6671e44a5d37c5c9ec5ebf