Properly handle OUT_DIR that can be relative to $TOP.
The build server calls gradle using $TOP as the cwd, not
framework/support. It also passes OUT_DIR as a relative path
to $TOP instead of an absolute path.

This made the Gradle script imcorrectly compute the true location
of out to be under frameworks/support.

In turn this leads to build server having output files stored
outside of their out/ folder causing impact on builds.

Bug: 18024010

Change-Id: Ie856d172bc74acb6cf827cfc2536be01f8b04d97