Material Dialog themes
Lots of changes needed for this, building upon the
refactoring CL. AppCompatDelegate can now handle Dialogs
as well as Activitys (which was easy after the refactor).

This was needed so that we can install a custom decor
for Dialogs.

Also added borderless button styles as they were
needed for Dialogs.

Change-Id: I88a7e4de5de85477982e401c832aeae732f81647