Fix crash on pasting text in a WebView
WindowDecorActionBar and Chrome's ContentViewCore each have their own ActionMode
reference. ActionModeImpl.finish() nulls WindowDecorActionBar's reference and
calls mCallback.onDestroyActionMode() to null ContentViewCore's reference. But
if the callback is deferred, there is a period when the ActionMode is finished
(and mCallback is null), but ContentViewCore doesn't know. ContentViewCore may
try to invalidate() the ActionMode, which will crash on the null mCallback.

Make ActionModeImpl more permissive so that calling invalidate() during this
period does nothing.

BUG: 18758329

Change-Id: I03b5be7ed60eaac79c55d0e12b461bff2b1053ed