Honor app:theme in Toolbar on Lollipop
With the change to use android:theme in AppCompat, there
was a small regression when using app:theme on Toolbar on L+.

This CL manually themifies the context so that we don't
break apps which only use app:theme when running on >= Lollipop.

Also added a Log statement telling devs that app:theme is deprecated.

Change-Id: Ib98f7c5b8b28d79c7b302a923fb198d8068035ca