Ignore hidden views for adapter helper
When adapter helper processes updates, it is important for it
whether a View is already in the UI or not. When we run
that check, we would usually check hidden views as well.

This becomes an issue if hidden view is not up to date because
it will be recycled while trying to re-use it in pre-layout.
(end animation callback will recycle it because it is not
up to date)

This CL changes AdapterHelper to ignore hidden views so that
they can continue their animation and LM will animate the rest.

Bug: 19170543
Change-Id: Ic26efd4e2a215481415131715625d785edc4f6ef