Handle sub view focus in RecyclerView
RecyclerView's child focus handler checks child's decor offsets
to ensure its decorations are also in viewport. There was a bug
in this section where if a descendant of an item view is focused,
it would crash. This CL fixes it by ignoring item decors if
focused view is not a RecyclerView child.

Bug: 19282711
Change-Id: I0f32a6d8ecfeea6f429f102b8d471251e278cd27