Gingerbread Item Animator bugfix
DefaultItemAnimator was adding view holder to the list
of remove animations after it starts the animation.

This was creating a bug because on Gingerbread, calling start
instantly calls end, which is responsible to remove the view
 holder. Since view holder is added after start, it was being
added after "end" is called as well, causing a leak. Also,
as a side effect, ItemAnimator was always thinking that it has
a running animation.

Change-Id: I80fecec93ff87f704383e4a26ba595b5c8edf1da