workaround for aar dependencies in pom.
Apparently depending on a non jar dependency in
a pom must include the type of the dependency, even though
the dependency declares its own type in its pom.

Since Gradle doesn't properly do this, a workaround
is to customize the pom on the fly to do this (until Gradle
is fixed).

Change-Id: Id39ae4b4183e621ba1295ae8ee840fd0125e5d6e