Fix RecyclerView focus issues when target view is too big
This CL fixes a bug where if the View which asked for focus is partially
visible, we would scroll such that the invisible part of the view also
becomes visible. Unfortunately, this scroll was not capped so if view is
too big, we would scroll too much and hide the currently visible side.

Bug: 19642142
Change-Id: I408b7c8aa544cde8d4de5050e080f5431aa3cd87