Fix IllegalArgumentException when using NUM_TRACKS or YEAR metadata on API14-19
MediaSessionCompatApi14 attempts to add metadata for the number of tracks and year to RemoteControlClient.MetadataEditor, causing an IllegalArgumentException as MetadataEditor does not support NUM_TRACKS at all and only supports YEAR as a long (not string) on API19. The only workaround available without this fix is to not use these fields at all if you want to support <API21 devices and use MediaSessionCompat.

BUG: 20144001
Change-Id: Iac3538d98a24489a9ca9689d9b13f6ff2287ea89