Make fragment host usable
There were some limitations in the API of the fragment host. This
should address them and allow the creation of a replacemnt fragment
host.
* added getHost() method; this is akin to the getActivity() method
and allows fragments to get a reference to their host. For the
FragmentActivity host, this method will return the same value as
getActivity()
* made all methods in FragmentController 'public'. This was an
oversight. There should be no package private methods in
FragmentController.
* removed "final" keyword from Fragment#getContext(). While this
method should be final, "getContext()" is too generic/common and
likely to be defined in existing Fragments. We're removing "final"
from the declaration to preven compile-time errors in legacy code.

Bug: 19569096
Change-Id: Id54e5febfdf2b2525c165ebb23f370f71f0ed0b8